+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Tests/Pages/TaskList/GeneralInformation/ViabilityIssuesIntegrationTests.cs
@@ -24,7 +24,7 @@ public class ViabilityIssuesIntegrationTests : BaseIntegrationTests
          x.Urn = project.Urn;
       });
 
-      await OpenAndConfirmPathAsync($"/task-list/{project.Id}/confirm-general-information");
+      await OpenAndConfirmPathAsync($"/task-list/{project.Id}/school-overview");
       await NavigateAsync("Change", 1);
 
       Document.Url.Should().BeUrl($"/task-list/{project.Id}/confirm-general-information/viability-issues");
@@ -39,7 +39,7 @@ public class ViabilityIssuesIntegrationTests : BaseIntegrationTests
 
       await Document.QuerySelector<IHtmlFormElement>("form")!.SubmitAsync();
 
-      Document.Url.Should().BeUrl($"/task-list/{project.Id}/confirm-general-information");
+      Document.Url.Should().BeUrl($"/task-list/{project.Id}/school-overview");
    }
 
    [Fact]
@@ -63,6 +63,6 @@ public class ViabilityIssuesIntegrationTests : BaseIntegrationTests
       await OpenAndConfirmPathAsync($"/task-list/{project.Id}/confirm-general-information/viability-issues");
       await NavigateAsync("Back");
 
-      Document.Url.Should().BeUrl($"/task-list/{project.Id}/confirm-general-information");
+      Document.Url.Should().BeUrl($"/task-list/{project.Id}/school-overview");
    }
 }